disable_publish_for_branches: new configuration option
OBS currently disables publishing for any new branch in order to avoid
using disk space and bandwidth to carry out publishing. For small private
installations, it can make sense to avoid that, and inherit the publish
flags from the parent project instead.

This is a backport to our 2.3.8-based OBS; the patches I've sent
upstream are rather different, because the way OBS deals with
configuration has changed significantly.

Bug: https://github.com/openSUSE/open-build-service/issues/596
Applied-upstream: no