cleanup_empty_projects: add new configuration option
Normally, if a submit request is accepted and results in cleaning up
the last package in a branched project, the branched project is
deleted too. This is annoying if the branched project is a long-running
branch (e.g. for a developer in a small team) and has been reconfigured,
for instance to turn on publishing in installations where branches do
not normally have publishing.

This is a backport to our 2.3.8-based OBS; the patches I've sent
upstream are rather different, because the way OBS deals with
configuration has changed significantly.

Bug: https://github.com/openSUSE/open-build-service/issues/596
Applied-upstream: no