Srdiff: Bail immediately if the tarlimit has been exceeded
Stop diffing tarballs if the line limit has been exceeded. All that
would be shown in that case is the number of lines that changed, but
that's not helpful if you can't see the diff. This makes diffing large
tarballs much faster in the common case where a tarlimit of 16000 is
used.

https://github.com/endlessm/eos-shell/issues/6225